requirements.txt
pandas>=1.3.0
numpy>=1.21.0
pyyaml>=5.4.1
scikit-learn>=1.0.0
openpyxl>=3.0.0
streamlit>=1.28.0
pyarrow>=5.0.0